Fixed issue number #2 from utilityCombatPlugin
And for some reason the project wasn't on Ue5. Lol.

One day I'll properly have this a fork of my own project, so things update cleanly. That day is not today. Ill have to test this throughly before I package my changes to the marketplace. I'll deal with it later as I am tired and i don't want to wait for changes to compile.